This commit was manufactured by cvs2svn to create tag 'RELEASE_1_0_1'.

+++ ext/sqlite/libsqlite/src/table.c
@@ -31,7 +31,7 @@ typedef struct TabResult {
   int nAlloc;
   int nRow;
   int nColumn;
-  long nData;
+  int nData;
   int rc;
 } TabResult;
